Main window: Move its utility functions to its own module
